Move tests in a better place
