+++ Sofa/Component/SolidMechanics/FEM/Elastic/tests/TetrahedronFEMForceField_test.cpp
@@ -20,6 +20,8 @@
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
 #include <sofa/component/solidmechanics/fem/elastic/TetrahedronFEMForceField.h>
+#include <sofa/simulation/common/SceneLoaderXML.h>
+
 #include "BaseTetrahedronFEMForceField_test.h"
 
 namespace sofa
@@ -49,6 +51,34 @@ public:
         stiffnessMat = tetraFEM->getMaterialStiffness(elementId);
         strainD = tetraFEM->getStrainDisplacement(elementId);
     }
+
+    void checkGracefullHandlingWhenTopologyIsMissing()
+    {
+        // This is a RAII message.
+        EXPECT_MSG_EMIT(Error) ;
+
+        std::stringstream scene ;
+        scene << R"scn(
+<?xml version='1.0'?>
+<Node name="root">
+    <RequiredPlugin name="Sofa.Component.StateContainer"/>
+    <RequiredPlugin name="Sofa.Component.SolidMechanics.FEM.Elastic"/>
+    <DefaultAnimationLoop/>
+    <Node name="FEMnode">
+        <MechanicalObject/>
+        <TetrahedronFEMForceField name="fem" youngModulus="5000" poissonRatio="0.07" printLog="true"/>
+    </Node>
+</Node>
+)scn";
+
+        m_root = simulation::SceneLoaderXML::loadFromMemory ("testscene", scene.str().c_str()) ;
+        m_root->init(sofa::core::execparams::defaultInstance());
+
+        core::objectmodel::BaseObject* fem = m_root->getTreeNode("FEMnode")->getObject("fem") ;
+        EXPECT_NE(fem, nullptr) ;
+
+        EXPECT_EQ(fem->getComponentState(), core::objectmodel::ComponentState::Invalid) ;
+    }
 };
 
 TEST_F(TetrahedronFEMForceField_test, init)
@@ -64,6 +94,11 @@ TEST_F(TetrahedronFEMForceField_test, FEMValues)
 TEST_F(TetrahedronFEMForceField_test, emptyTology)
 {
     this->checkEmptyTopology();
+}
+
+TEST_F(TetrahedronFEMForceField_test, checkGracefullHandlingWhenTopologyIsMissing)
+{
+    this->checkGracefullHandlingWhenTopologyIsMissing();
 }
 
 } // namespace sofa
